Dev/leo

+++ frontend/src/utils/converter.util.ts
@@ -4,7 +4,7 @@ import { Schedule } from "@/types/schedule.type"
 export const formatDate = (isoString: string, withTime: boolean = false): string => {
     const date = new Date(isoString)
     const day = date.getDate()
-    const month = date.toLocaleString("en-US", { month: "long" })
+    const month = date.toLocaleString("en-US", { month: "short" })
     const year = date.getFullYear()
   
     if (!withTime) return `${day} ${month} ${year}`
